Update external-prs app to namespace
