+++ ops/external-prs/package.json
@@ -5,6 +5,9 @@
   "author": "Kariba Labs",
   "license": "Apache-2.0",
   "private": true,
+  "bin": {
+    "external-prs": "./dist/src/cli.js"
+  },
   "main": "./dist/src/index.js",
   "types": "./dist/src/index.d.ts",
   "type": "module",
@@ -31,12 +34,17 @@
   },
   "dependencies": {
     "@google-cloud/bigquery": "^7.5.1",
+    "@types/columnify": "^1.5.4",
     "@types/libsodium-wrappers": "^0.7.13",
     "@types/yargs": "^17.0.32",
     "chalk": "^5.3.0",
+    "columnify": "^1.6.0",
     "dayjs": "^1.11.10",
+    "duckdb": "^0.10.1",
     "libsodium-wrappers": "^0.7.13",
     "octokit": "^3.1.0",
+    "oss-directory": "^0.0.10",
+    "tmp-promise": "^3.0.3",
     "ts-dedent": "^2.2.0",
     "winston": "^3.11.0",
     "yaml": "^2.3.1",
