+++ ops/external-prs/src/base.ts
@@ -1,0 +1,9 @@
+import { Repo } from "./github.js";
+import { App } from "octokit";
+
+export interface BaseArgs {
+  githubAppPrivateKey: string;
+  githubAppId: string;
+  repo: Repo;
+  app: App;
+}
