+++ ops/external-prs/test.out
@@ -1,0 +1,5 @@
+deploy=true
+sha=ba596ee5d4adeb728365d106dc3392e2becb689a
+pr=1071
+issue_author=oso-lets-go
+comment_author=ravenac95
